Use Convert.ToByte
